+++ brain-react/constants/data.jsx
@@ -1,0 +1,7 @@
+export class INIT_TODO_LIST {
+
+}
+
+export class INIT_UNIQUE_ID {
+
+}
